refactor: move topology cleanup to optimize/, reroot to treetime_graph
- Move reroot types (`RerootChanges`, `EdgeSplitInfo`, `EdgeMergeInfo`, `RerootResult`) and functions (`split_edge`, `apply_reroot_topology`, `remove_node_if_trivial`) to `treetime_graph::reroot` - pure graph-key structs and graph operations with no domain dependencies
- Move `collapse`, `merge_shared_mutations`, `polytomy_nodes` from `partition/algo/topology_cleanup/` to `optimize/topology/` - tree cleanup operations used during optimization, not partition behavior
- Remove empty `partition/algo/topology_cleanup/` directory
- Update all imports across `clock/`, `timetree/`, `optimize/`, `commands/`, and test files
