+++ packages/treetime-graph/src/lib.rs
@@ -6,3 +6,4 @@ pub mod graph;
 pub mod graph_ops;
 pub mod graph_traverse;
 pub mod node;
+pub mod reroot;
