+++ packages/treetime/src/clock/__tests__/test_reroot.rs
@@ -5,7 +5,7 @@ mod tests {
   use crate::clock::find_best_root::params::BranchPointOptimizationParams;
   use crate::clock::reroot::{RerootParams, reroot_in_place};
   use crate::o;
-  use crate::partition::algo::topology_cleanup::reroot::remove_node_if_trivial;
+  use treetime_graph::reroot::remove_node_if_trivial;
   use eyre::Report;
   use maplit::btreemap;
   use pretty_assertions::assert_eq;
