+++ packages/treetime/src/commands/prune/run.rs
@@ -3,8 +3,8 @@ use crate::ancestral::fitch::create_fitch_partition;
 use crate::commands::prune::args::TreetimePruneArgs;
 use crate::gtr::get_gtr::{GtrModelName, get_gtr_by_name, log_gtr, write_gtr_json};
 use crate::make_error;
-use crate::partition::algo::topology_cleanup::collapse::collapse_edge;
-use crate::partition::algo::topology_cleanup::merge_shared_mutations::merge_shared_mutation_branches;
+use crate::optimize::topology::collapse::collapse_edge;
+use crate::optimize::topology::merge_shared_mutations::merge_shared_mutation_branches;
 use crate::partition::marginal_dense::PartitionMarginalDense;
 use crate::partition::marginal_sparse::PartitionMarginalSparse;
 use crate::partition::payload::ancestral::GraphAncestral;
