+++ packages/treetime/src/optimize/run_loop.rs
@@ -4,8 +4,8 @@ use crate::optimize::dispatch::run_optimize_mixed_inner;
 use crate::optimize::indel::{estimate_indel_rate, total_indel_log_lh};
 use crate::optimize::iteration::{apply_damping, restore_branch_lengths, save_branch_lengths};
 use crate::optimize::params::{BranchOptMethod, InitialGuessMode};
-use crate::partition::algo::topology_cleanup::collapse::collapse_edge;
-use crate::partition::algo::topology_cleanup::merge_shared_mutations::merge_shared_mutation_branches;
+use crate::optimize::topology::collapse::collapse_edge;
+use crate::optimize::topology::merge_shared_mutations::merge_shared_mutation_branches;
 use crate::partition::marginal_dense::PartitionMarginalDense;
 use crate::partition::marginal_sparse::PartitionMarginalSparse;
 use crate::partition::traits::{HasGtr, PartitionOptimizeOps, PartitionOptimizeVec};
