+++ packages/treetime/src/optimize/topology/collapse.rs
