+++ packages/treetime/src/optimize/topology/mod.rs
@@ -4,4 +4,3 @@ mod __tests__;
 pub mod collapse;
 pub mod merge_shared_mutations;
 pub mod polytomy_nodes;
-pub mod reroot;
