+++ packages/treetime/src/partition/algo/mod.rs
@@ -1,2 +1,1 @@
 pub mod infer_dense;
-pub mod topology_cleanup;
